Redirecter alltid til skjemaoversikt ved 404. Refactor for kodegjenbruk i 404 og redirect wrapper.
