+++ src/sider/404/404.tsx
@@ -1,15 +1,8 @@
 import React, { Component, CSSProperties } from "react";
-import { Sidetittel, Normaltekst } from "nav-frontend-typografi";
-import { Link } from "react-router-dom";
-import { FormattedMessage } from "react-intl";
-import errorIcon from "../../img/error.png";
 import { injectIntl, InjectedIntlProps } from "react-intl";
-import { sideTittel, storForsteBokstav } from "../../utils/sprak";
-import Helmet from "react-helmet";
-import { getApplicableSkjemaoversiktRedirect } from "../../redirects/redirects";
 import { RouteComponentProps, withRouter } from "react-router";
-import { AllRoutes } from "../../redirects/NySkjemaoversiktRedirects";
-import Spinner from "../../komponenter/spinner/Spinner";
+import { AllRoutes } from "../../redirects/MedSkjemaoversiktRedirects";
+import { RedirectTilSkjemaoversikt } from "../../redirects/RedirectTilSkjemaoversikt";
 
 type Props = {
   style?: CSSProperties;
@@ -19,53 +12,7 @@ type Props = {
 
 class NotFound extends Component<Props> {
   render() {
-    const { intl, match, style, message } = this.props;
-
-    if (intl.locale !== "en") {
-      const redirectUrl = getApplicableSkjemaoversiktRedirect(match.params);
-
-      window.location.replace(redirectUrl);
-
-      return <Spinner />;
-    }
-
-    const title = sideTittel(storForsteBokstav(`${intl.formatMessage({ id: "notFound.tittel" })}`));
-
-    return (
-      <div className="notFound__container" style={style}>
-        <Helmet>
-          <title>{title}</title>
-        </Helmet>
-        <img className="notFound__icon" alt={message} src={errorIcon} />
-        <div className="notFound__title">
-          <Sidetittel>
-            <FormattedMessage id="notFound.tittel" />
-          </Sidetittel>
-        </div>
-        {message && (
-          <div className="notFound__message">
-            <Normaltekst>
-              <FormattedMessage id={message} />
-            </Normaltekst>
-          </div>
-        )}
-        <div className="notFound__content">
-          <Normaltekst>
-            <FormattedMessage id="notFound.beskrivelse" />
-          </Normaltekst>
-        </div>
-        <Normaltekst>
-          <a href="https://nav.no">
-            <FormattedMessage id="notFound.goto.nav" />
-          </a>
-        </Normaltekst>
-        <Normaltekst>
-          <Link to="/soknader">
-            <FormattedMessage id="notFound.goto.soknadsveiviser" />
-          </Link>
-        </Normaltekst>
-      </div>
-    );
+    return <RedirectTilSkjemaoversikt params={this.props.match.params} />;
   }
 }
